Restructure folders for a cleaner server.js
