+++ server.js
@@ -6,13 +6,11 @@ const mongoose = require('mongoose');
 const express = require('express');
 const cors = require('cors');
 const passport = require('passport');
-const passportLocal = require('passport-local').Strategy;
 const cookieParser = require('cookie-parser');
-const bcrypt = require('bcryptjs');
 const session = require('express-session');
 const bodyParser = require('body-parser');
 const app = express();
-const User = require('./user');
+
 mongoose.connect(
   "mongodb+srv://Michael:b88bs@pamslist.zeeew.mongodb.net/test?retryWrites=true&w=majority",
   {
@@ -23,7 +21,10 @@ mongoose.connect(
     console.log("Mongoose is connected")
   }
 );
+
+// create instance of socket server 
 const server = require('http').createServer(app);
+// enable CORS for local dev
 const io = require('socket.io')(server, {
   cors: {
     origin: "http://localhost:3000",
@@ -54,16 +55,19 @@ require('./passportConfig')(passport);
 
 //---------------End of Middleware-------------------
 
+//---------------POSTGRES STUFF---------------------- 
 // sync server to db
-const db = require("./models");
-db.sequelize.sync();
+// const db = require("./models");
+// db.sequelize.sync();
 
 // test route
-app.get('/', (req, res) => {
-  res.send('Hello World!');
-});
+// app.get('/', (req, res) => {
+//   res.send('Hello World!');
+// });
 
-// socket.io events
+//---------------SOCKET IO---------------------------
+
+// events
 io.on('connection', (socket) => {
   // let addedUser = false;
 
@@ -83,48 +87,12 @@ io.on('connection', (socket) => {
 //--------------------------------------------------
 //----------------------ROUTES----------------------
 //--------------------------------------------------
-app.post('/login', (req, res, next) => {
-  passport.authenticate('local', (err, user, info) => {
-    if (err) throw err;
-    if (!user) res.send('No user exits!');
-    else {
-      req.logIn(user, err => {
-        if (err) throw err;
-        res.send('Successfully Authenticated');
-        console.log(req.user);
-      })
-    }
-  })(req, res, next);
-});
 
-app.post('/register', (req, res) => {
-  User.findOne({ username: req.body.username }, async (err, doc) => {
-    if (err) throw err;
-    if (doc) res.send("User already exists");
-    if (!doc) {
-      const hashedPassword = await bcrypt.hash(req.body.password, 10);
+const userRoutes = require('./routes/user');
+const testRoute = require('./routes/test');
 
-      const newUser = new User({
-        username: req.body.username,
-        password: hashedPassword
-      });
-      await newUser.save();
-      res.send("User Created");
-    }
-  })
-});
-
-app.get('/user', (req, res) => {
-  res.send(req.user);
-});
-
-app.get('/logout', (req, res) => {
-  console.log('logging out')
-  req.logout();
-  res.redirect('/');
-});
-
-
+app.use('/', userRoutes);
+app.use('/', testRoute);
 
 //--------------------------------------------------
 //-------------------Start Server-------------------
